\name{FunctionsBCR-package}
\alias{FunctionsBCR-package}
\alias{FunctionsBCR}
\docType{package}
\title{General utility functions for Bird Conservancy of the Rockies}
\description{General purpose utility functions for science and data operations at Bird Conservancy of the Rockies}
\author{Quresh Latif}
\keyword{ package }
