\name{gatherNimble}
\alias{gatherNimble}
\title{
  Gather NIMBLE samples
}
\description{
  Base function called by `RunNimbleParallel` used to gather and organize NIMBLE samples into an mcmcOutput object.
}
\usage{
  gatherNimble(read.path, burnin = 1)
}
\arguments{
  \item{read.path}{Path to directory containing sample blocks generated by `RunNimbleParallel`.}
  \item{burnin}{Number of sampling blocks to discard at the beginning of each chain when gathering MCMC samples from saved blocks.}
}
\value{MCMC samples organized as an 'mcmcOutput' object.}
\author{
  Bryan L. Nuse and Quresh S. Latif, Bird Conservancy of the Rockies
}
\keyword{nimble}
\keyword{mcmcOutput}
\keyword{RunNimbleParallel}