\name{RunNimbleParallel}
\alias{RunNimbleParallel}
\title{
  NIMBLE wrapper with parallel processing and automated sampling until convergence
}
\description{
  Implements Bayesian model fitting using the NIMBLE R packaged in parallelized clusters across multiple processors. Additionally, convergence and sampling metrics (i.e., Rhat and n_effective) are checked periodically, and sampling continues until these metrics meet specified minimum levels. Saved samples are periodically dumped to the hard drive to avoid overloading RAM with successive rounds of resuming sampling.
  
  ***Note: This function relies on submitting Linux commands via the processx R package, and thus will not operate in Windows.
}
\usage{
  RunNimbleParallel(model, inits, data, constants, parameters,
           par.ignore.Rht = c(), par.dontign.Rht = c(),
           nc = 2, ni = 2000, nb = 0.5, nt = 10, mod.nam = "mod",
           max.samples.saved = 10000, rtrn.model = F, sav.model = T,
           Rht.required = 1.1, neff.required = 100, check.freq = 60,
           max.tries = NULL)
}
\arguments{
  \item{model}{Object containing NIMBLE model code generated by the function nimble::nimbleCode.}
  \item{inits}{Function that generates initial values for stochastic nodes as required to run a NIMBLE model.}
  \item{data}{List containing all data objects required to fit specified NIMBLE model.}
  \item{constants}{List containing all constants required to fit specified NIMBLE model.}
  \item{parameters}{String vector containing names of all parameters for which estimates are to be stored by the sampler.}
  \item{par.ignore.Rht}{String vector listing parameters to be ignored when calculating convergence and sampling metrics. All parameters with names containing strings listed here will be ignored. Default is an empty vector, whereby all parameters will be considered for calculating Rhat.}
  \item{par.dontign.Rht}{String vector listing parameters that should not be ignored when calculating convergence and sampling metrics. All parameters with names containing strings listed here will not be ignored. This parameter can be used in conjunction with 'par.ignore.Rht' to focus convergence checking on desired parameters. Default is an empty vector, whereby no parameters will be designated as not to be ignored for calculating Rhat.}
  \item{nc}{Number of MCMC chains to run. Note available cores must equal or exceed 'nc'. Default = 2.}
  \item{ni}{Number of iterations to run MCMC chains before checking convergence and sampling metrics. Default = 10000. If model does not meet specified levels for R_hat and n_effective, sampling will continue for another 'ni' steps before checking metrics again.}
  \item{nb}{If < 1, proportion of MCMC chains to discard as burn-in. If > 1, chain length before thinning to discard as burn-in.}
  \item{nt}{Base level of thinning to implement within parallelized cluster. Default = 10. Setting this argument will reduce load on RAM for large models.}
  \item{mod.nam}{Character string used as a file label for saved model outputs.}
  \item{max.samples.saved}{Integer that sets max samples to save across chains in final model object. Additional thinning of posterior samples (i.e., after samples are recovered from processing clusters) is implemented as necessary to limit the size of the model object to this level. Default = 10000. For no cap, set to NULL.}
  \item{rtrn.model}{Logical indicating whether to return model output to R environment. Default = F, i.e., model output is only saved to disk.}
  \item{sav.model}{Logical indicating whether to save model output to disk. Default = T, i.e., model output is saved to disk.}
  \item{Rht.required}{Maximum numeric value for Rhat for any parameter required to end sampling.}
  \item{neff.required}{Minimum value for n_effective for any parameter required to end sampling.}
  \item{check.freq}{How long to wait in seconds between successive checks of convergence criteria.}
  \item{max.tries}{The maximum number of times to check convergence, after which RunNimbleParallel will quit warning. Default is to not set any constraint on number of checks.}
}
\value{If assigned to object in global environment and rtrn.model = T, model output is returned. Else, model output is saved to file. Regardless, returned value consists of a list with three elements:
  \item{mcmcOutput}{posterior samples as returned by mcmcOutput::mcmcOutput}
  \item{summary}{data frame containing summaries of posterior estimates and convergence criteria (Rhat and n_effective) for all parameters in mcmcOutput}
  \item{mcmc.info}{A vector with elemets: nchains = number of MCMC sampling chains, niterations = pre-thinning chain length including burnin, burnin = pre-thinning length of each chain discarded as burn-in, nthin = realized thinning expressed as the ith sample retained. Note that nthin can be a non-whole number if max.samples.saved is specified, in which case it is the mean interval between retained samples.}
  }
\author{
  Quresh S. Latif and Bryan L. Nuse, Bird Conservancy of the Rockies
}

\keyword{nimble}
\keyword{mcmcOutput}
\keyword{parallel processing}
